requirements.txt
google-generativeai>=0.3.0
numpy>=1.21.0
pandas>=1.3.0